Fixes None bug in os.path.isfile call
In the _cleanUp method of VideoConverter class, arguments
were not check to see if valid (i.e., not None). This caused
exception in call to os.path.isfile. This has been fixed by
checking if arg and os.path.isfile(arg), which will fail before
the os.path.isfile() is reached in the case that arg in None.

v1.2.1

+++ video_utils/videoconverter.py
@@ -398,7 +398,7 @@ class VideoConverter( ComRemove, MediaInfo, OpenSubtitles ):
   def _cleanUp(self, *args):
     ''' Method to delete arbitrary number of files, catching exceptions'''
     for arg in args:
-      if os.path.isfile( arg ):
+      if arg and os.path.isfile( arg ):
         try:
           os.remove( arg )
         except Excpetion as err:
